+++ test/test034.cpp
@@ -1,0 +1,11 @@
+#include "../iostream.h" 
+#include "../shed_math.h"  
+#include "../shed_types.h"
+
+int main(){
+    shed_std::Cconsole_output << shed_std::INT64_MAX << shed_std::end_line;
+    shed_std::Cconsole_output << shed_std::INT64_MIN << shed_std::end_line;
+
+    shed_std::Cconsole_output << shed_std::MATH_PI << shed_std::end_line;
+    shed_std::Cconsole_output << shed_std::MATH_E  << shed_std::end_line;
+}